requirements.txt
aioagi==1.2.0
mautrix>=0.19.16,<=0.20.0
ruamel.yaml>=0.15.35,<0.18
jinja2>=2,<4
jinja2-ansible-filters==1.3.2 
